# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nimiq Mini Apps API",
    "version": "1.0.0",
    "description": "REST API for the [Nimiq Mini Apps](https://nimiqminiapps.com) community directory.\n\n**Submitting an app** — `POST /api/apps/submit` (wallet login required). Rate limit: **5 requests per hour per IP**.\nRequired field `submitter_contact` (Telegram, email, etc.) is stored privately and omitted from public responses.\n\n**Open in Nimiq Pay** — listed apps expose `open_url` as `https://nimpay.app/miniapps/open/{domain}` (domain without `https://`).\n\nMachine-readable spec: `GET /openapi.json` or `GET /openapi.yaml`.\n",
    "contact": {
      "name": "Nimiq Mini Apps",
      "url": "https://nimiqminiapps.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/NimMiniApps/NimiqMiniApps/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://api.nimiqminiapps.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/apps": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List apps",
        "operationId": "listApps",
        "parameters": [
          {
            "$ref": "#/components/parameters/query"
          },
          {
            "$ref": "#/components/parameters/category"
          },
          {
            "$ref": "#/components/parameters/developer"
          },
          {
            "$ref": "#/components/parameters/tag"
          },
          {
            "$ref": "#/components/parameters/asset"
          },
          {
            "$ref": "#/components/parameters/rewards"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/featured"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/collection"
          },
          {
            "$ref": "#/components/parameters/paginate"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "description": "App list, or paginated result when `paginate=1`",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AppPublic"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedApps"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/apps/{slug}/related": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Related apps (same developer or category)",
        "operationId": "getRelatedApps",
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Up to 4 related public apps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppPublic"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/developers/{slug}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get developer and their public apps",
        "operationId": "getDeveloper",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Developer profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Developer"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/admin/apps": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all apps (any status)",
        "operationId": "adminListApps",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full catalog including submitted and rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppAdmin"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create app",
        "operationId": "adminCreateApp",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppAdmin"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Set ADMIN_TOKEN on the server; send as Authorization Bearer token"
      },
      "walletCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wallet_session",
        "description": "Signed wallet-session cookie set by POST /api/auth/verify"
      }
    },
    "parameters": {
...
      "query": {
        "name": "q",
        "in": "query",
        "description": "Search name, tagline, description, tags, assets, reward assets, developer",
        "schema": {
          "type": "string"
        }
      },
      "category": {
        "name": "category",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/CategoryName"
        }
      },
      "developer": {
        "name": "developer",
        "in": "query",
        "description": "Filter by developer_slug",
        "schema": {
          "type": "string"
        }
      },
      "tag": {
        "name": "tag",
        "in": "query",
        "description": "Exact tag match",
        "schema": {
          "type": "string"
        }
      },
      "asset": {
        "name": "asset",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/Asset"
        }
      },
      "rewards": {
        "name": "rewards",
        "in": "query",
        "description": "Use `true` to return only apps with one or more reward assets.",
        "schema": {
          "type": "boolean"
        }
      },
      "status": {
        "name": "status",
        "in": "query",
        "description": "Public listings default to approved, verified, experimental",
        "schema": {
          "$ref": "#/components/schemas/AppStatus"
        }
      },
      "featured": {
        "name": "featured",
        "in": "query",
        "schema": {
          "type": "boolean"
        }
      },
      "sort": {
        "name": "sort",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "featured",
            "newest",
            "name"
          ],
          "default": "featured"
        }
      },
      "collection": {
        "name": "collection",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "new-week",
            "popular",
            "rewards",
            "games",
            "usdt"
          ]
        }
      },
      "paginate": {
        "name": "paginate",
        "in": "query",
        "schema": {
          "type": "boolean"
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100
        }
      },
      "offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AppPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryName"
          },
          "developer_slug": {
            "type": "string"
          },
          "developer_name": {
            "type": "string"
          },
          "owner_wallet_addresses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Wallets that can manage this listing (My apps, request-update). Empty when unclaimed."
          },
          "tagline": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "long_description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "reward_assets": {
            "type": "array",
            "description": "Assets users can actually receive from this app, such as daily rewards, leaderboard prizes, payouts, tips, faucets, or similar receive-side flows. Leave empty when the app merely accepts, displays, swaps, or supports a token.",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "status": {
            "$ref": "#/components/schemas/AppStatus"
          },
          "release_stage": {
            "$ref": "#/components/schemas/ReleaseStage"
          },
          "featured": {
            "type": "boolean"
          },
          "featured_order": {
            "type": "integer"
          },
          "website_url": {
            "type": "string",
            "nullable": true
          },
          "github_url": {
            "type": "string",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "nullable": true
          },
          "discovered_icon_url": {
            "type": "string",
            "nullable": true
          },
          "banner_url": {
            "type": "string",
            "nullable": true
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItem"
            }
          },
          "socials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLink"
            }
          },
          "domain_reachable": {
            "type": "boolean",
            "nullable": true
          },
          "domain_checked_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "open_url": {
            "type": "string",
            "format": "uri",
            "example": "https://nimpay.app/miniapps/open/myapp.example.com"
          }
        }
      },
...
    }
  }
}